+++ packages/server/lib/modes/index.ts
@@ -10,6 +10,8 @@ export = (mode, options) => {
     return require('./smoke_test').run(options)
   }
 
+  options.isBrowserGivenByCli = options.browser !== undefined
+
   if (mode === 'run') {
     _.defaults(options, {
       socketId: random.id(10),