+++ system-tests/projects/config-default-browser/cypress.config.js
@@ -1,0 +1,8 @@
+const { defineConfig } = require('cypress')
+
+module.exports = defineConfig({
+  e2e: {
+    supportFile: false,
+  },
+  defaultBrowser: 'chrome',
+})